requirements.txt
fastapi
uvicorn
websockets
jinja2
python-multipart
deepgram-sdk>=5.0.0
groq
edge-tts
python-dotenv